Zero first-party compiler warnings; MultiDelay heading overlap fixed
Casts, renames and unused-code removals only, no behaviour change. The
MultiDelay GLOBAL heading is now anchored below the Pan label at every
window size, and MultiDelay joins the UI snapshot gate.

+++ HyperPrismReimagined/Source/Compressor/CompressorEditor.cpp
@@ -173,7 +173,7 @@ void GainReductionMeter::paint(juce::Graphics& g)
         if (db % 10 == 0)
         {
             g.setFont(10.0f);
-            g.drawText(juce::String(db), x - 5, bounds.getY() + 7, 10, 20, 
+            g.drawText(juce::String(db), static_cast<int>(x - 5), static_cast<int>(bounds.getY() + 7), 10, 20,
                       juce::Justification::centred);
         }
     }
